requirements.txt
pandas>=2.2.2,<3.0.0
numpy>=2.1.0,<3.0.0
scikit-learn>=1.5.1
flask==3.0.3
werkzeug==3.0.3
plotly==5.23.0
dash==2.17.1
pyyaml==6.0.2
python-dotenv==1.0.1
joblib==1.4.2
uvicorn==0.30.6
orjson==3.10.7
streamlit>=1.28.0,<2.0.0
